+++ routes/campuses/drivers-positions.mjs
@@ -3,11 +3,14 @@ import maskOutput, { cleanObject } from '../../middlewares/mask-output';
 
 import GeoTracking from '../../models/geo-tracking';
 import Campus from '../../models/campus';
+import resolveRights from '../../middlewares/check-rights';
+import { CAN_GET_RIDE_POSITION } from '../../models/rights';
 
 const router = new Router();
 
 router.get(
   '/',
+  resolveRights(CAN_GET_RIDE_POSITION),
   maskOutput,
   async (ctx) => {
     const drivers = await Campus.findDrivers(ctx.params.campus_id);
@@ -21,4 +24,4 @@ router.get(
   },
 );
 
-export default router.routes();
+export default router;