+++ routes/phones.mjs
@@ -19,6 +19,9 @@ const router = generateCRUD(Phone, {
   },
   list: {
     right: CAN_LIST_PHONE_LOCAL,
+    filters: {
+      campus: 'campus._id',
+    },
     middlewares: [
       contentNegociation,
       maskOutput,
@@ -62,7 +65,7 @@ const router = generateCRUD(Phone, {
   },
   batch: {
     right: CAN_CREATE_PHONE_LOCAL,
-    refs: ['_id', 'label'],
+    refs: ['_id', 'imei', 'number'],
     middlewares: [
       csvToJson,
       validateCampus,
@@ -70,4 +73,4 @@ const router = generateCRUD(Phone, {
   },
 });
 
-export default router.routes();
+export default router;